move: StatusIcon and PackageCount under generic components
Previously the components were placed under ContentListTable.
Since they are only presentational components
and used under Templates and Repositories pages both,
they are now placed under generic components.
