+++ src/Pages/Templates/TemplatesTable/components/AddOrEditTemplate/steps/CustomRepositoriesStep.tsx
@@ -27,8 +27,8 @@ import Hide from 'components/Hide/Hide';
 import { SkeletonTable } from '@patternfly/react-component-groups';
 import { Table, TableVariant, Tbody, Td, Th, ThProps, Thead, Tr } from '@patternfly/react-table';
 import UrlWithExternalIcon from 'components/UrlWithLinkIcon/UrlWithLinkIcon';
-import PackageCount from 'Pages/Repositories/ContentListTable/components/PackageCount';
-import StatusIcon from 'Pages/Repositories/ContentListTable/components/StatusIcon';
+import PackageCount from 'components/PackageCount/PackageCount';
+import StatusIcon from 'components/StatusIcon/StatusIcon';
 import useDebounce from 'Hooks/useDebounce';
 import { ADD_ROUTE, REPOSITORIES_ROUTE } from 'Routes/constants';
 import TdWithTooltip from 'components/TdWithTooltip/TdWithTooltip';
